% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBC.support.code.R
\docType{package}
\name{ACBC-package}
\alias{ACBC-package}
\alias{ACBC}
\title{Analysis of convex body coverage}
\value{
Key functions for this package:
\item{\code{\link{achc}}}{Permutation procedure to generate many random convex hulls and measure
the amount of space covered by individual hulls.}
\item{\code{\link{aec}}}{Permutation procedure to generate many random ellipsoids and measure
the amount of space covered by individual ellipsoids.}
\item{\code{\link{plot.achc}}}{Generates plots associated with achc.}
\item{\code{\link{plot.aec}}}{Generates plots associated with aec}
The name, "ACBC", is an acronym for, "Analysis of Convex Body Coverage."  Through
the various functions in this package, one can determine if the distribution of convex bodies in a data
space, represented by groups to be compared, is consistent with or different than expectation from
random assembly of observations.  ACBC fits a uniform grid of points to the total expanse of multidimensional space
covered by observations.  Convex hulls or ellipsoids for groups are described, and the frequency of 0, 1, ..., g convex bodies for 
g groups are measured at every point.  A relative frequency distribution is generated from these frequencies.  Next,
the procedure is repeated many times with random assigment of observations to groups.  Confidence bands
are generated for the 0, 1, 2, ..., g groups.  By comparing the observed distribution to the confidence limits
generated, one can ascertain if the data space is meaningfully partitioned in some way.
}
\author{
Michael Collyer
}
