% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACHC.support.code.R
\docType{package}
\name{ACHC-package}
\alias{ACHC-package}
\alias{ACHC}
\title{Analysis of convex hull coverage}
\value{
Key functions for this package:
\item{\code{\link{achc}}}{Permutation procedure to generate many random hulls and measure
the amount of space covered by individual hulls.}
\item{\code{\link{plot.achc}}}{Generates plots associated with ACHC.}
The name, "ACHC", is an acronym for, "Analysis of Convex Hull Coverage."  Through
the various functions in this package, one can determine if the distribution of convex hulls in a data
space, represented by groups to be compared, is consistent with or different than expectation from
random assembly of observations.  ACHC fits a uniform grid of points to the total expanse of multidimensional space
covered by observations.  Hulls for groups are described, and the frequency of 0, 1, ..., g hulls for 
g groups are measured at every point.  A relative frequency distribution is generated from these frequencies.  Next,
the procedure is repeated many times with random assigment of observations to groups.  Confidence bands
are generated for the 0, 1, 2, ..., g groups.  By comapring the observed distribution to the confidence limits
generated, one can ascertain if the data space is meaningfully partitioned in some way.
}
\author{
Michael Collyer
}
