% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achc.R
\name{achc}
\alias{achc}
\title{Analysis of Convex Hull Coverage}
\usage{
achc(dat, group, grid.points = 500, grid.space = 0.05, iter = 99,
  seed = NULL, print.progress = TRUE)
}
\arguments{
\item{dat}{Either a data frame or matrix of data to be analyzed.  Because of computational
limitations, the number of variables is currently limited to 5.}

\item{group}{A factor or vector coerible to factor for defining groups.}

\item{grid.points}{The desired number of points, sampled from a uniform distribution of points
in the data space, within a convex hull for all observed points.  This number might be
less than the maximum possible number of points (which could be huge).}

\item{grid.space}{The approximate spacing of uniform points along each axis.  For example, 0.05 means points will
be placed at increments that are 5 percent of the expanse of data, per axis.}

\item{iter}{The number of iterations (permutations) to run for the test.  Because the observed case counts as one
iteration, this should be the number desired, minus one.}

\item{seed}{AChange the random seed, if desired.  If NULL, the seed will equal the number of permutations.}

\item{print.progress}{A logical value to indicate if permutation progress should be printed to the screen.
This is useful for analyses that will run a long time.}
}
\value{
An object of class \code{achc} is a list containing the following
\item{grid}{grid points obtained.}
\item{analysis}{A matrix of 0s and 1s for convex hull presenece at grid points (rows)
by groups (columns) for each permutation.}
\item{group}{The factor of group levels used; useful for downstream functions.}
}
\description{
Function obtains relative frequencies of convex hull coverage in the represented data space,
and uses a permutation procedure to generate confidence limits for random assignmnet of observations to groups.
}
\details{
A description is needed here
}
\examples{

# 3 dimensions of data
library(RRPP)
data("Pupfish")
group <- interaction(Pupfish$Sex, Pupfish$Pop)
P <- prcomp(Pupfish$coords)$x[,1:3] # first 3 PCs

grid.preview(P, pts = 500, pt.scale = 0.05)
pupCHC <- achc(P, group, iter = 99, grid.points = 500, grid.space = 0.05)
pupCHC
summary(pupCHC, confidence = 0.95)
plot(pupCHC, lwd = 2)

# The grid used
library(rgl)
plot3d(pupCHC$grid)
aspect3d("iso")

# Example of 8-dimensional data analysis
data(PupfishHeads)
group <- factor(paste(PupfishHeads$locality, PupfishHeads$year, sep = "."))
P <- prcomp(PupfishHeads$coords)$x[, 1:8] # first 8 PCs

grid.preview(P, pts = 500, pt.scale = 0.05)
pupCHC <- achc(P, group, iter = 99, grid.points = 500, grid.space = 0.05)
pupCHC
summary(pupCHC, confidence = 0.95)
plot(pupCHC, lwd = 2)

}
\author{
Michael Collyer
}
\keyword{analysis}
