% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.preview.R
\name{grid.preview}
\alias{grid.preview}
\title{Preview of Uniform points}
\usage{
grid.preview(Y, pts = 500, pt.scale = 0.05, ...)
}
\arguments{
\item{Y}{A matrix or data frame of data}

\item{pts}{The number of points (in all dimensions)}

\item{pt.scale}{The relative distance between points, as a fraction of the expanse 
of representation of data (by variable).  the default, 0.05, means a 5 percent
increment along axes.}

\item{...}{other arguments for graphical parameters}
}
\description{
Function generates a pairs plot to show the hypothetical distribution of 
uniform points that might be generated for \code{\link{achc}}.
}
\examples{
library(RRPP)
data("Pupfish")
P <- prcomp(Pupfish$coords)$x[,1:3] # first 3 PCs

grid.preview(P, pts = 100, pt.scale = 0.1, pch = 19, cex = 0.3, asp = 1)
grid.preview(P, pts = 500, pt.scale = 0.05, pch = 19, cex = 0.3, asp = 1)
}
\author{
Michael Collyer
}
\keyword{utilities}
