% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACHC.utils.R
\name{plot.achc}
\alias{plot.achc}
\title{Plot Function for ACHC}
\usage{
\method{plot}{achc}(x, ...)
}
\arguments{
\item{x}{plot object (from \code{\link{achc}})}

\item{...}{other arguments passed to plot (helpful to employ
different colors or symbols for different groups).  See
\code{\link{plot.default}} and \code{\link{par}}.  Some limitations
occur because of polygon use.}
}
\description{
Plot Function for ACHC
}
\author{
Michael Collyer
}
\keyword{utilities}
\keyword{visualization}
