% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBC.utils.R
\name{print.achc}
\alias{print.achc}
\title{Print/Summary Function for ACHC}
\usage{
\method{print}{achc}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{achc}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for ACHC
}
\author{
Michael Collyer
}
\keyword{utilities}
