% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACHC.utils.R
\name{print.aec}
\alias{print.aec}
\title{Print/Summary Function for AEC}
\usage{
\method{print}{aec}(x, ...)
}
\arguments{
\item{x}{print/summary object (from \code{\link{achc}})}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for AEC
}
\author{
Michael Collyer
}
\keyword{utilities}
