% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACHC.utils.R
\name{summary.achc}
\alias{summary.achc}
\title{Print/Summary Function for ACHC}
\usage{
\method{summary}{achc}(object, confidence = 0.95, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{achc}})}

\item{confidence}{Confidence level for ACHC test}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for ACHC
}
\author{
Michael Collyer
}
\keyword{utilities}
