% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBC.utils.R
\name{summary.aec}
\alias{summary.aec}
\title{Print/Summary Function for AEC}
\usage{
\method{summary}{aec}(object, confidence = 0.95, ...)
}
\arguments{
\item{object}{print/summary object (from \code{\link{achc}})}

\item{confidence}{Confidence level for AEC test}

\item{...}{other arguments passed to print/summary}
}
\description{
Print/Summary Function for AEC
}
\author{
Michael Collyer
}
\keyword{utilities}
