% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achc.R
\name{achc}
\alias{achc}
\title{Analysis of Convex Hull Coverage}
\usage{
achc(dat, group, grid.points = 500, iter = 99, seed = NULL,
  print.progress = TRUE)
}
\arguments{
\item{dat}{Either a data frame or matrix of data to be analyzed.  Because of computational
limitations, the number of variables is currently limited to 5.}

\item{group}{A factor or vector coerible to factor for defining groups.}

\item{grid.points}{The targeted number of uniform points to use to defined the expanse of the data space.  Emphasis 
is placed on maintaining uniformity, so the optimal solution will likely not match exactly the desired number of points.
Optimization involves fitting a spectrum of point desnities and choosing the solution that most closely matches the 
number of grid points chosen.  Disparity between targeted and obtained numbers are likely due to edge effects
(as small changes in density can add or remove many landmarks at the edges of data space occupation).}

\item{iter}{The number of iterations (permutations) to run for the test.  Because the observed case counts as one
iteration, this should be the number desired, minus one.}

\item{seed}{AChange the random seed, if desired.  If NULL, the seed will equal the number of permutations.}

\item{print.progress}{A logical value to indicate if permutation progress should be printed to the screen.
This is useful for analyses that will run a long time.}
}
\value{
An object of class \code{achc} is a list containing the following
\item{grid}{grid points obtained.}
\item{analysis}{A matrix of 0s and 1s for convex hull presenece at grid points (rows)
by groups (columns) for each permutation.}
\item{group}{The factor of group levels used; useful for downstream functions.}
}
\description{
Function obtains relative frequencies of convex hull coverage in the represented data space,
and uses a permutation procedure to generate confidence limits for random assignmnet of observations to groups.
}
\details{
A description is needed here
}
\examples{

library(RRPP)
data("Pupfish")
group <- interaction(Pupfish$Sex, Pupfish$Pop)
P <- prcomp(Pupfish$coords)$x[,1:3] # first 3 PCs

pupCHC <- achc(P, group, iter = 99, grid.points = 1000)
pupCHC
summary(pupCHC, confidence = 0.95)
plot(pupCHC, lwd = 2)

# The grid used
library(rgl)
plot3d(pupCHC$grid)
aspect3d("iso")

}
\author{
Michael Collyer
}
\keyword{analysis}
